requirements.txt
asgiref==3.8.1
Django==4.2.18
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
